refactor of documents
